refactor(command-visualizer): split pipeline.ts under 250-line limit
Extract two cohesive pure steps out of fixtureToCascade:
- derive-flashes.ts: step 6b char-diff synthesis of pendingDelete/justAdded
  when a fixture records no ide.flashes but the doc changed.
- derive-overlays.ts: step 7 highlights + thatMark/sourceMark -> decorations.
Both return decoration lists the caller appends; order and behavior identical.
Removed now-unused Pos/pos imports. Rendered output byte-identical.
